refactor(coding-agent): extract agent-session types module
Move the public session type vocabulary (AgentSessionConfig, PromptOptions,
HandoffResult, model-cycle/stats/advisor result types, FreshSessionResult, and
the internal ActiveAdvisor/AdvisorRuntimeDescriptor) into sibling
agent-session-types.ts. Pure type move; the old ./agent-session path re-exports
the full public surface, so importers are unchanged. Types are erased, so the
runtime export surface is byte-identical.

Op: compress
