+++ packages/coding-agent/src/session/agent-session.ts
@@ -84,7 +84,6 @@ import type {
 	Context,
 	ImageContent,
 	Message,
-	MessageAttribution,
 	Model,
 	ProviderResponseMetadata,
 	ProviderSessionState,
@@ -313,6 +312,23 @@ import { normalizeModelContextImages } from "../utils/image-loading";
 import { describeAttachedImagesForTextModel } from "../utils/image-vision-fallback";
 import { generateSessionTitle } from "../utils/title-generator";
 import { buildNamedToolChoice, isToolChoiceActive } from "../utils/tool-choice";
+import type {
+	ActiveAdvisor,
+	AdvisorRuntimeDescriptor,
+	AdvisorStats,
+	AgentSessionConfig,
+	ContextUsageBreakdown,
+	FreshSessionResult,
+	HandoffResult,
+	ModelCycleResult,
+	PerAdvisorStat,
+	PromptOptions,
+	ResolvedRoleModel,
+	RoleModelCycle,
+	RoleModelCycleResult,
+	SessionHandoffOptions,
+	SessionStats,
+} from "./agent-session-types";
 import type { AuthStorage } from "./auth-storage";
 import type { ClientBridge, ClientBridgePermissionOption, ClientBridgePermissionOutcome } from "./client-bridge";
 import {
@@ -635,359 +651,24 @@ export interface AsyncJobSnapshot {
 	delivery: AsyncJobDeliveryState;
 }
 
+// Public session types now live in ./agent-session-types (T14). Re-exported
+// here so the old ./agent-session path keeps its full surface.
+export type {
+	AdvisorStats,
+	AgentSessionConfig,
+	ContextUsageBreakdown,
+	FreshSessionResult,
+	HandoffResult,
+	ModelCycleResult,
+	PerAdvisorStat,
+	PromptOptions,
+	ResolvedRoleModel,
+	RoleModelCycle,
+	RoleModelCycleResult,
+	SessionHandoffOptions,
+	SessionStats,
+} from "./agent-session-types";
 export type { ShakeMode, ShakeResult };
-
-// ============================================================================
-// Types
-// ============================================================================
-
-export interface AgentSessionConfig {
-	agent: Agent;
-	sessionManager: SessionManager;
-	settings: Settings;
-	/** Whether the caller explicitly requested yolo/auto-approve behavior for this session. */
-	autoApprove?: boolean;
-	/** Models to cycle through with Ctrl+P (from --models flag) */
-	scopedModels?: Array<{ model: Model; thinkingLevel?: ThinkingLevel }>;
-	/** Initial session thinking selector. */
-	thinkingLevel?: ConfiguredThinkingLevel;
-	/** Initial per-family service tiers (OpenAI / Anthropic / Google) for the live session. */
-	serviceTierByFamily?: ServiceTierByFamily;
-	/** Prompt templates for expansion */
-	promptTemplates?: PromptTemplate[];
-	/** File-based slash commands for expansion */
-	slashCommands?: FileSlashCommand[];
-	/** Extension runner (created in main.ts with wrapped tools) */
-	extensionRunner?: ExtensionRunner;
-	/** Loaded skills (already discovered by SDK) */
-	skills?: Skill[];
-	/** Skill loading warnings (already captured by SDK) */
-	skillWarnings?: SkillWarning[];
-	/** Custom commands (TypeScript slash commands) */
-	customCommands?: LoadedCustomCommand[];
-	skillsSettings?: SkillsSettings;
-	/** Model registry for API key resolution and model discovery */
-	modelRegistry: ModelRegistry;
-	/** Tool registry for LSP and settings */
-	toolRegistry?: Map<string, AgentTool>;
-	/** Tool names whose current registry entry is still the built-in implementation. */
-	builtInToolNames?: Iterable<string>;
-	/** Current session pre-LLM message transform pipeline */
-	transformContext?: (messages: AgentMessage[], signal?: AbortSignal) => AgentMessage[] | Promise<AgentMessage[]>;
-	/**
-	 * Per-request transform applied after `convertToLlm` and before the
-	 * provider call. Used for snapcompact, secret obfuscation, and image
-	 * clamping. When supplied via {@link createAgentSession}, the advisor agent
-	 * inherits this so its requests undergo the same shaping as the main turn.
-	 */
-	transformProviderContext?: (context: Context, model: Model) => Context | Promise<Context>;
-	/**
-	 * Stream wrapper passed to side-channel requests (`/btw`, `/omfg`, IRC
-	 * auto-replies, and handoff generation) so they apply the same provider
-	 * shaping and host-level request wrappers as normal agent turns. Defaults
-	 * to plain `streamSimple` when omitted.
-	 */
-	sideStreamFn?: StreamFn;
-	/**
-	 * Stream wrapper passed to the advisor agent so its requests apply the
-	 * session's `providers.openrouterVariant`, `providers.antigravityEndpoint`,
-	 * `providers.maxInFlightRequests`, and `model.loopGuard.*` settings —
-	 * keeping OpenRouter sticky-routing / response caching consistent with the
-	 * main agent. Defaults to plain `streamSimple` when omitted.
-	 */
-	advisorStreamFn?: StreamFn;
-	/** Hint that OpenAI Codex requests should prefer websocket transport when supported. */
-	preferWebsockets?: boolean;
-	/** Provider payload hook used by the active session request path */
-	onPayload?: SimpleStreamOptions["onPayload"];
-	/** Provider response hook used by the active session request path */
-	onResponse?: SimpleStreamOptions["onResponse"];
-	/** Raw SSE hook used by the active session request path */
-	onSseEvent?: SimpleStreamOptions["onSseEvent"];
-	/** Per-session raw SSE diagnostic buffer */
-	rawSseDebugBuffer?: RawSseDebugBuffer;
-	/** Current session message-to-LLM conversion pipeline */
-	convertToLlm?: (messages: AgentMessage[]) => Message[] | Promise<Message[]>;
-	/** System prompt builder that can consider tool availability. Returns ordered provider-facing blocks. */
-	rebuildSystemPrompt?: (toolNames: string[], tools: Map<string, AgentTool>) => Promise<{ systemPrompt: string[] }>;
-	/** Rebuild the SSH tool from current capability discovery results. */
-	reloadSshTool?: () => Promise<AgentTool | null>;
-	requestedToolNames?: ReadonlySet<string>;
-	/**
-	 * Optional accessor for live MCP server instructions. Read by the session's
-	 * `rebuildSystemPrompt`-skip optimization to detect server-side instruction
-	 * changes (e.g. an MCP server upgrade) that would otherwise pass the tool-set
-	 * signature comparison and silently keep a stale prompt cached.
-	 */
-	getMcpServerInstructions?: () => Map<string, string> | undefined;
-	/** Enable hidden-by-default MCP tool discovery for this session. */
-	mcpDiscoveryEnabled?: boolean;
-	/** MCP tool names to activate for the current session when discovery mode is enabled. */
-	initialSelectedMCPToolNames?: string[];
-	/** Whether constructor-provided MCP defaults should be persisted immediately. */
-	persistInitialMCPToolSelection?: boolean;
-	/** MCP server names whose tools should seed discovery-mode sessions whenever those servers are connected. */
-	defaultSelectedMCPServerNames?: string[];
-	/** MCP tool names that should seed brand-new sessions created from this AgentSession. */
-	defaultSelectedMCPToolNames?: string[];
-	/** TTSR manager for time-traveling stream rules */
-	ttsrManager?: TtsrManager;
-	/** Secret obfuscator for deobfuscating streaming edit content */
-	obfuscator?: SecretObfuscator;
-	/** Inherited eval executor session id from a parent agent. */
-	parentEvalSessionId?: string;
-	/** Logical owner for retained eval kernels created by this session. */
-	evalKernelOwnerId?: string;
-	/**
-	 * AsyncJobManager that this session installed as the process-global instance.
-	 * Only set for top-level sessions; subagents inherit the parent's manager and
-	 * **MUST NOT** dispose it on their own teardown.
-	 */
-	ownedAsyncJobManager?: AsyncJobManager;
-	/**
-	 * AsyncJobManager reachable by this session for scoped job actions.
-	 *
-	 * Top-level owners receive their own manager, subagents receive the inherited
-	 * parent manager, and secondary in-process top-level sessions receive
-	 * `undefined` so job snapshots and ACP drains cannot observe the primary's
-	 * state.
-	 */
-	asyncJobManager?: AsyncJobManager;
-	/** Agent identity (registry id like "Main" or "Alice") used for IRC routing. */
-	agentId?: string;
-	/** Whether this session is the top-level agent or a subagent. Drives eager-task
-	 *  prelude gating so a top-level session created with a custom `agentId` still
-	 *  receives the always-mode reminder. Defaults to "main". */
-	agentKind?: "main" | "sub";
-	/**
-	 * Override the provider-facing session ID for all API requests from this session.
-	 * When absent, `sessionManager.getSessionId()` is used. Needed when benchmark or
-	 * SDK callers issue probes / prewarming with an explicit `--provider-session-id`
-	 * so that credential sticky selection is consistent with the session's streaming calls.
-	 */
-	providerSessionId?: string;
-	/**
-	 * Full advisor toolset, pre-built in `createAgentSession` against a distinct,
-	 * advisor-scoped `ToolSession` (its own `-advisor` session/agent id) so the
-	 * advisor's tool state stays isolated from the primary. The advisor is a full
-	 * agent; its config `tools` selects a subset (default read/grep/glob). Undefined
-	 * when the advisor is disabled.
-	 */
-	advisorTools?: AgentTool[];
-	/** Preloaded watchdog prompt content for the advisor. */
-	advisorWatchdogPrompt?: string;
-	/** Preloaded YAML top-level `instructions` shared baseline, kept separate from
-	 *  `advisorWatchdogPrompt` so `/advisor configure` can swap it live. */
-	advisorSharedInstructions?: string;
-	/**
-	 * Preloaded project context files (AGENTS.md, etc.) rendered as a system-prompt
-	 * block for the advisor — the same standing instructions the primary agent
-	 * receives, so the reviewer holds the agent to them.
-	 */
-	advisorContextPrompt?: string;
-	/**
-	 * Advisors discovered from `WATCHDOG.yml`. Empty/undefined runs a single
-	 * legacy advisor on the `advisor` role (byte-for-byte the pre-config path).
-	 */
-	advisorConfigs?: AdvisorConfig[];
-	/**
-	 * Strip tool descriptions from provider-bound tool specs on side requests
-	 * (handoff). Must match the session-start value used to build the system
-	 * prompt so inline descriptors are not also sent through provider schemas.
-	 */
-	pruneToolDescriptions?: boolean;
-	/**
-	 * Disconnect this session's OWNED MCP manager on dispose. Provided only when
-	 * the session created the manager (top-level sessions); subagents reuse a
-	 * parent's manager via `options.mcpManager` and omit this so a child's
-	 * teardown never tears down the shared servers.
-	 */
-	disconnectOwnedMcpManager?: () => Promise<void>;
-	/**
-	 * Override the bundled system prompt used by automatic session-title
-	 * generation paths (initial title + replan refresh). Source-of-truth is
-	 * `TITLE_SYSTEM.md` discovered via {@link discoverTitleSystemPromptFile} and
-	 * resolved through {@link resolvePromptInput}; refresh after a `/move`-style
-	 * cwd change via {@link AgentSession.setTitleSystemPrompt}.
-	 */
-	titleSystemPrompt?: string;
-}
-
-/** Options for AgentSession.prompt() */
-export interface PromptOptions {
-	/** Whether to expand file-based prompt templates (default: true) */
-	expandPromptTemplates?: boolean;
-	/** Image attachments */
-	images?: ImageContent[];
-	/** When streaming, how to queue the message: "steer" (interrupt) or "followUp" (wait). */
-	streamingBehavior?: "steer" | "followUp";
-	/** Optional tool choice override for the next LLM call. */
-	toolChoice?: ToolChoice;
-	/** Send as developer/system message instead of user. Providers that support it use the developer role; others fall back to user. */
-	synthetic?: boolean;
-	/** Marks this prompt as a deliberate user action (typed message, `.`/`c`
-	 *  continue). Clears advisor auto-resume suppression that a user interrupt set.
-	 *  Defaults to `!synthetic`; manual-continue is synthetic yet user-initiated, so
-	 *  it sets this explicitly. Agent-initiated synthetic prompts (auto-continue,
-	 *  plan re-prime, reminders) leave it unset and keep suppression latched. */
-	userInitiated?: boolean;
-	/** Explicit billing/initiator attribution for the prompt. Defaults to user prompts as `user` and synthetic prompts as `agent`. */
-	attribution?: MessageAttribution;
-	/** Skip pre-send compaction checks for this prompt (internal use for maintenance flows). */
-	skipCompactionCheck?: boolean;
-}
-
-/** Result from a handoff operation. */
-export interface HandoffResult {
-	document: string;
-	savedPath?: string;
-}
-
-export interface SessionHandoffOptions {
-	autoTriggered?: boolean;
-	signal?: AbortSignal;
-}
-
-/** Result from cycleModel() */
-export interface ModelCycleResult {
-	model: Model;
-	thinkingLevel: ThinkingLevel | undefined;
-	/** Whether cycling through scoped models (--models flag) or all available */
-	isScoped: boolean;
-}
-
-/** Result from cycleRoleModels() */
-export interface RoleModelCycleResult {
-	model: Model;
-	thinkingLevel: ThinkingLevel | undefined;
-	role: string;
-}
-
-/** A configured role resolved to a concrete model, used by role cycling and
- *  the plan-approval model slider. */
-export interface ResolvedRoleModel {
-	role: string;
-	model: Model;
-	thinkingLevel?: ConfiguredThinkingLevel;
-	explicitThinkingLevel: boolean;
-}
-
-/** The set of resolvable role models plus the index of the currently active
- *  one within {@link ResolvedRoleModel.role} order. */
-export interface RoleModelCycle {
-	models: ResolvedRoleModel[];
-	currentIndex: number;
-}
-
-export interface ContextUsageBreakdown {
-	contextWindow: number;
-	anchored: boolean;
-	usedTokens: number;
-	systemPromptTokens: number;
-	systemToolsTokens: number;
-	systemContextTokens: number;
-	skillsTokens: number;
-	messagesTokens: number;
-}
-
-/** Session statistics for /session command */
-export interface SessionStats {
-	sessionFile: string | undefined;
-	sessionId: string;
-	userMessages: number;
-	assistantMessages: number;
-	toolCalls: number;
-	toolResults: number;
-	totalMessages: number;
-	tokens: {
-		input: number;
-		output: number;
-		reasoning: number;
-		cacheRead: number;
-		cacheWrite: number;
-		total: number;
-	};
-	premiumRequests: number;
-	cost: number;
-}
-
-/** Advisor statistics for /advisor status command. */
-export interface AdvisorStats {
-	configured: boolean;
-	active: boolean;
-	model?: Model;
-	contextWindow: number;
-	contextTokens: number;
-	tokens: {
-		input: number;
-		output: number;
-		reasoning: number;
-		cacheRead: number;
-		cacheWrite: number;
-		total: number;
-	};
-	cost: number;
-	messages: {
-		user: number;
-		assistant: number;
-		total: number;
-	};
-	/** Per-advisor breakdown; one entry per active advisor (single-advisor sessions have one). */
-	advisors: PerAdvisorStat[];
-}
-
-/** One advisor's slice of {@link AdvisorStats}, surfaced for the multi-advisor status panel. */
-export interface PerAdvisorStat {
-	name: string;
-	model: Model;
-	contextWindow: number;
-	contextTokens: number;
-	tokens: AdvisorStats["tokens"];
-	cost: number;
-	messages: AdvisorStats["messages"];
-}
-
-/**
- * One live advisor instance: its own agent/runtime/tools/recorder plus a
- * per-advisor emission guard and identity. The session holds an array of these;
- * primary-scoped state (turn counters, interrupt latches, the shared yield
- * channel) stays on the session.
- */
-interface ActiveAdvisor {
-	/** Display name from config ("default" for the legacy no-YAML advisor). */
-	name: string;
-	/** Slug for the transcript filename/session id; "" → `__advisor.jsonl`. */
-	slug: string;
-	agent: Agent;
-	runtime: AdvisorRuntime;
-	adviseTool: AdviseTool;
-	emissionGuard: AdvisorEmissionGuard;
-	recorder: AdvisorTranscriptRecorder;
-	/** Latest recorder close, awaited by dispose() so the final turn lands on disk. */
-	recorderClosed: Promise<void>;
-	/** Unsubscribe for the advisor agent's event stream feeding the recorder. */
-	agentUnsubscribe?: () => void;
-	model: Model;
-	thinkingLevel: ThinkingLevel;
-	/** Stable key for the resolved runtime inputs that require a rebuild to change. */
-	signature: string;
-}
-
-/** Resolved advisor config ready to instantiate as an {@link ActiveAdvisor}. */
-interface AdvisorRuntimeDescriptor {
-	config: AdvisorConfig;
-	name: string;
-	slug: string;
-	model: Model;
-	thinkingLevel: ThinkingLevel;
-	signature: string;
-}
-
-export interface FreshSessionResult {
-	previousSessionId: string;
-	sessionId: string;
-	closedProviderSessions: number;
-}
 
 /** Internal marker for hook messages queued through the agent loop */
 // ============================================================================
